Created a main GameScreen
The Asteroids class now handles the different screens in the game
